# docs/engineering/openapi.yaml
openapi: 3.1.0
info:
  title: OpsFlow API
  version: current-v1
  license:
    name: MIT
    identifier: MIT
  description: |
    OpenAPI summary for the currently implemented Express API.
    This file is intentionally concise; see api-design.md for the human-readable route notes.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
...
paths:
...
  /api/jobs/{jobId}/evidence:
    get:
      tags: [Job]
      operationId: listJobEvidence
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          $ref: "#/components/responses/PagelessListSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Job]
      operationId: uploadJobEvidence
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        $ref: "#/components/requestBodies/UploadEvidence"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/jobs/{jobId}/evidence/{evidenceId}:
    delete:
      tags: [Job]
      operationId: deleteJobEvidence
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/EvidenceId"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    EvidenceId:
      name: evidenceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    UploadEvidence:
      required: true
      content:
        multipart/form-data:
          schema:
            $ref: "#/components/schemas/UploadEvidenceRequest"
...
  responses:
    Success:
      description: Successful JSON response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiSuccessResponse"
    PagelessListSuccess:
      description: Successful JSON response with an array data payload.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiSuccessResponse"
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
...
    BadRequest:
      description: Validation failed or request payload is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    NotFound:
      description: Resource does not exist or is not visible in the current tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
...
    JobEvidenceKind:
      type: string
      enum: [SITE_PHOTO, COMPLETION_PROOF, CUSTOMER_DOCUMENT, ISSUE_EVIDENCE]
...
    UploadEvidenceRequest:
      type: object
      required: [kind, file]
      properties:
        kind:
          $ref: "#/components/schemas/JobEvidenceKind"
        note:
          type: string
          maxLength: 300
        file:
          type: string
          format: binary
...
    ApiSuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data: {}
        meta:
          type: object
          additionalProperties: true
    ApiErrorResponse:
      type: object
      required: [success, code, message]
      properties:
        success:
          type: boolean
          enum: [false]
        code:
          type: string
          description: Stable application error code for client handling and support workflows.
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true